Separate reward computation from reward function definition (make future swap to genesys easy)
